+++ js/util.js
@@ -1,0 +1,19 @@
+const idGenerator = function*(maxCount) {
+  let currentId = 1;
+  while (currentId <= maxCount) {
+    yield currentId++;
+  }
+};
+
+const getRandomInt = (min, max)=> {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
+
+const getRandomElementFromArray = (elementsArray) => {
+  const randomElement = elementsArray[getRandomInt(0, elementsArray.length - 1)];
+  return randomElement;
+};
+
+export {idGenerator, getRandomInt, getRandomElementFromArray};
